--- Chapter4/01-learningNpm/index.js
@@ -1,8 +1,0 @@
-const dateConvert = require('date-convert')
-const data = {
-    name: 'khanh',
-    birthday: '2000-02-17',
-    age: '20'
-}
-console.log(dateConvert(data.birthday))
-
