--- Chapter4/01-learningNpm/package.json
@@ -1,14 +1,0 @@
-{
-  "name": "learningnpm",
-  "version": "1.0.0",
-  "description": "npm learning",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "author": "Hoang Duc Khanh",
-  "license": "ISC",
-  "dependencies": {
-    "date-convert": "^0.1.2"
-  }
-}
