--- Chapter4/02-createBasicApi/.gitignore
@@ -1,2 +1,0 @@
-node_modules
-node_modules/
